Stuff be working, no bugs so far, time to implement stuff in the generator

+++ RootTools/AbilityEditor/Include/AbilityEditor.h
@@ -39,7 +39,7 @@ private slots:
 		void BrowseName();
 		void ChangeAbilityName();
 		void SetCooldown(double p_cooldown);
-		void SetCharges(double p_cooldown);
+		void SetCharges(int p_cooldown);
 		void SetChargeTime(double p_cooldown);
 		void SetChannelingTime(double p_cooldown);
 		void SetDuration(double p_duration);
